@@ -20,16 +20,16 @@
         "optimize-autoloader": true
     },
     "require":{
-        "php": "~7.0",
+        "php": "~7.4",
         "psr/http-message":"^1.0.0",
-        "teknoo/immutable": "^0.0.1"
+        "teknoo/immutable": "~2.0.0-beta5"
     },
     "require-dev": {
         "phpunit/phpunit": "~8.4",
         "squizlabs/php_codesniffer": "~3.5",
         "pdepend/pdepend": "~2.5",
-        "sebastian/phpcpd": "~3.0||~4.1",
-        "phploc/phploc": "~4.0||~5.0",
+        "sebastian/phpcpd": "~4.1",
+        "phploc/phploc": "~5.0",
         "phpstan/phpstan": "~0.10",
         "guzzlehttp/guzzle":"^6.0.0",
         "guzzlehttp/psr7":"^1.0.0",
