@@ -1,5 +1,25 @@
 <?php
 
+/*
+ * Sellsy Client.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license and the version 3 of the GPL3
+ * license that are bundled with this package in the folder licences
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richarddeloge@gmail.com so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) 2009-2016 Richard Déloge (richarddeloge@gmail.com)
+ *
+ * @link        http://teknoo.software/sellsy-client Project website
+ *
+ * @license     http://teknoo.software/sellsy-client/license/mit         MIT License
+ * @author      Richard Déloge <richarddeloge@gmail.com>
+ */
+
 namespace Teknoo\Sellsy\Definitions;
 
 use Teknoo\Sellsy\Client\ClientInterface;
@@ -16,9 +36,6 @@ use Teknoo\Sellsy\Method\Method;
  */
 class ElectronicSign implements DefinitionInterface
 {
-    /**
-     * @inheritdoc
-     */
     public function __invoke(ClientInterface $client): CollectionInterface
     {
         $collection = new Collection($client, 'ElectronicSign');
