@@ -1,6 +1,6 @@
 <?php
 
-/**
+/*
  * Sellsy Client.
  *
  * LICENSE
@@ -41,10 +41,6 @@ interface DefinitionInterface
 {
     /**
      * To create on the flow a new collection instance, regrouping several sellsy's methods.
-     *
-     * @param ClientInterface $client
-     *
-     * @return CollectionInterface
      */
     public function __invoke(ClientInterface $client): CollectionInterface;
 }
